requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
dlib==19.24.6
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
pillow==11.0.0
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
